Project TG Implementando comandos ultrapassar e relatorioUltrapassagens

+++ Models/Car.php
@@ -13,6 +13,13 @@ class Car
 
     public function newCar($pilot, $make, $model, $color, $year)
     {
+        $race = json_decode(file_get_contents(__DIR__ . '/../filesJson/dataRace.json'), true);
+
+        if ($race['Start'] == true) {
+            echo "A corrida foi Iniciada, voce nao pode adicionar carros" . PHP_EOL;
+            exit;
+        }
+
         $cars = json_decode($this->dataCars);
         unlink(__DIR__ . "/../filesJson/dataCars.json");
 
@@ -58,16 +65,21 @@ class Car
     {
         $cars = json_decode($this->dataCars, true);
 
-        foreach ($cars as $key=>$car) {
+        foreach ($cars as $key => $car) {
             $key = $key + 1;
             echo "-----------------------------" . PHP_EOL
-                . "Carro n ". $key . PHP_EOL
+                . "Carro n " . $key . PHP_EOL
                 . "Piloto - " . $car['Piloto'] . PHP_EOL
                 . "Marca - " . $car['Marca'] . PHP_EOL
                 . "Modelo - " . $car['Modelo'] . PHP_EOL
                 . "Cor - " . $car['Cor'] . PHP_EOL
-                . "Ano - " . $car['Ano'] . PHP_EOL
-                . "Posicao - " . $car['Posicao'] . PHP_EOL . PHP_EOL;
+                . "Ano - " . $car['Ano'] . PHP_EOL;
+
+            if (!empty($car['Posicao'])) {
+                echo "Posicao - " . $car['Posicao'] . PHP_EOL . PHP_EOL;
+            } else {
+                echo PHP_EOL;
+            }
         }
     }
 }
